run Byron server even if Shelley client is on
The cardano-sl diffusion layer offers no distinction between a Byron
server and client: if you want one but not the other, you still need to
bring the whole thing up. In Byron proxy at the moment, we can't deal
with more than one source for blocks, so in case a Shelley block source
is given, the Byron server can still be brought up, but blocks will only
be downlaoded from Shelley. The latest tip is announced every 10 seconds
to Byron peers.