Transaction submission protocol
This overhauls the node to node transaction submission protocol. The
new design is designed for full pipelining and to support what the
higher level transaction relaying logic needs to do. In particular to
enable forwarding dependent transactions in order without stalling we
need to be able to request transactions out of order, while still
limiting the amount of state that both peers need to track.

This patch covers the protocol type, client and server wrappers, the
direct connect check, examples and tests.

The higher level relaying logic remains to do.